UI: Fix disconnected and non-clickable checkbox labels in popovers
Follow the logic of
ed870f87b942: Fix weird placement of "Motion Tracking" in 3D View
overlay popup

But with an additional check, if this sub-panel is used in Popover,
and not inside the Sidebar or Property editor.

For panels, the checkbox inside panel header must not have a label, so
the panel can be collapsed/expanded, panels already have their own
label in the header, this remains unchanged with this patch.

For popovers, draw the checkbox in header as usual with label, but
hide the panel header label, see Julian's commit.

Patch by Yevgeny Makarov (@jenkm) in
https://archive.blender.org/developer/D9586