Sculpt: Added vector displacement for sculpting draw brush (area plane method only for now)
Vector displacement maps (VDM) provide a way to create complex displacements that can have overhangs in one brush dab.
This is unlike standard displacement with height maps that only displace in normal direction.
Forms like ears, curled horns, etc can be created in one click if VMDs are used.

This patch reads the RGB channels of a texture in a brush stroke and interprets them as individual vectors, that are used for the vertex offset.
As of now, this is only working for the draw brush using the area plane method. Symmetry and radial symmetry is working.