Constraints: use a dimensionless unit prop for Maintain Volume
The Volume property of this constraint is currently marked as a
distance, which makes it confusing--especially with non-metric units.

For example, try creating a 1-meter bone. Set its target volume to 1,
and its scale to 4. The constrained dimensions are computed as 0.5
(4.0 * 0.5 * 0.5 = 1.0). Now, try setting the units to imperial. The
target volume is converted to 3.28' and nothing makes sense any more.

The volume can actually be understood as a factor of the initial
volume, so it should be dimensionless.

Additionally, the volume had a range of 0.001 to 100.0. This is wide
enough in most cases, but sometimes you may need to go orders of
magnitude higher or lower to consider vast or thin objects, and there
should be no drawback to extending the limits, provided they stay
positive.