Subdivision Surface: add dependency graph tracking when cpu mesh is needed.
After addition of GPU subdivision it is necessary to choose between
using CPU or GPU for a particular object when both ways are technically
possible. Choosing GPU when the mesh is also needed by CPU code would
cause subdivision to be run twice when the CPU mesh is requested by
other code.

This patch adds a simple way to track dependencies that need the CPU
mesh through dependency graph via an already existing mechanism.

Constraints targeting mesh vertex groups, as well as likely modifiers
now tag their targets with a dependency graph flag to notify that the
fully evaluated mesh is needed on the CPU. This is used to disable GPU
subdiv in those cases.

In addition, the evaluated mesh is used by sculpt and paint modes,
so GPU subdivision is disabled in that case too.