Cycles: Merge Anisotropic BSDF node into Glossy BSDF node
Internally these are already using the same code path anyways, there's no point
in maintaining two distinct nodes.

The obvious approach would be to add Anisotropy controls to the Glossy BSDF node
and remove the Anisotropic BSDF node. However, that would break forward
compability, since older Blender versions don't know how to handle the Anisotropy
input on the Glossy BSDF node.

Therefore, this commit technically removes the Glossy BSDF node, uses versioning
to replace them with an Anisotropic BSDF node, and renames that node to "Glossy
BSDF".

That way, when you open a new file in an older version, all the nodes show up as
Anisotropic BSDF nodes and render correctly.

This is a bit ugly internally since we need to preserve the old `idname` which
now no longer matches the UI name, but that's not too bad.