Cycles: Use per-microfacet Fresnel term for Glass closures
This commit replaces the current Glass approach, where Glass is a virtual closure
that gets replaced with a Glossy and a Refractive closure, with a combined
closure that handles Fresnel after sampling the microfacet. That way, the Fresnel
term is more accurate since it accounts for the microfacet normal, not the
shading normal.

Two TODOs:
- OSL support is still missing, should be easy to add
- The code needs a third random variable for picking reflection or refraction.
  Currently it just hashes `randu` and `randv`, but that breaks stratification
  and significantly increases noise. The best option appears to be introducing
  a second 2D sample pair for BSDFs - that way, we can use it here for a nicely
  stratified sample, and we can even get rid of the LCG in the principled hair
  closure (since that requires 4 dimensions). That way, once we kick out
  MultiGGX, we can remove the shader LCG entirely.