MSVC: lower C4457 warning level from 4 to 3
This commit lowers the warning level for C44571, which occurs when a
local variable hides a function parameter with the same name.

The corresponding diagnostic options for GCC and Clang is `-Wshadow`,
which is curretly enabled for Blender.

By lowering the warning level for C4457, we align the warnings with
other compilers and avoid unnecessary noise.

Note that in MSVC, the warning only checks the name, whereas the
corresponding one in other compilers apparently checks the
[name + type] match. Then some warnings popped up and the code needed
to be edited.