MSVC: lower C4100 warning level from 4 to 3
This better aligns with OSX/Linux warnings.

---
Although `__pragma(warning(suppress:4100))` is not the same as `__attribute__((__unused__))` in gcc (which only affects the attribute instead of the line), it still seems to be better to use it than to hide the warning entirely.

So, after lowering the warning level, adding this `pragma` to the existing `UNUSED` definitions suppressed most of the warnings.

In some projects, such as those in extern, this warning is suppressed via the cmake macros `remove_strict_flags()` and `remove_extra_strict_flags()`. So to handle those in MSVC too, it was enough to add `remove_cc_flag(/w34100)` to these macros.

Other cases the warning was specially suppressed for "GNUC" and "Clang" but missing for "MSVC". In these cases the corresponding MSVC pragma has been added.

Lastly, two projects had to be handled differently in MSVC because the warning appears in the windows libraries. These projects are `Alembic` and `USD`.

(Note that some C++ files used only for windows have been edited to silence the warning).