[Sequencer] Don't make undo items when an empty sequencer gets clicked
If the sequencer receives a click event, it looks for items to
select/deselect.

When the sequencer is empty (i.e., there are no sequences), we would have
the `deselect_all` variable set to `true` called
`ED_sequencer_deselect_all` to select any existing sequences.

The code just assumed that because the `deselect_all` flag was true and we
called `ED_sequencer_deselect_all` that something changed. That's not true
when the sequencer is empty.

Changes made
------------
I've changed `ED_sequencer_deselect_all` to return a boolean value
indicating whether or not it has mark any sequences for deselection.

The `changed` variable is set to the value returned by
`ED_sequencer_deselect_all`.

When there are no sequences present, `changed` is false.

Notes
-----

I noticed that if I click and drag (Box selection) around the editor even
when there are no sequences an empty Box selection undo item gets created.

We can either update the parent task to get that fixed or create a new
task.

I'm gonna deal with that after this diff.

Test plan:

1. Check-out this diff
2. Build blender
3. Open the Sequencer
4. Go to Edit > Undo History and confirm the history is empty
5. Click all over the Sequencer window
6. Go to Edit > Undo History and confirm the history is still empty

Differential Revision: https://developer.blender.org/D17212