Mesh: Set active attribute values edit mode operator
This patch adds a simple operator to set values of the active attribute for
the selected element. The aim is to give simple control over attribute values
in edit mode rather than to provide the fastest workflow for most cases.
Eventually this operator might be less important compared to more advanced
attribute editing tools, but for now, exposing a little bit of functionality
is low hanging fruit and will help to see the possibilities.

The implementation mostly consists of boilerplate to register the necessary
property types for the operator and draw their UI. Beyond that, we just loop
over selected elements and set a value.