Geometry Nodes: Edges to Face Groups Node
Add a new node that groups faces inside of boundary edge regions.
This is the opposite action as the existing "Face Group Boundaries" node[1].
It's also the same as some of the "Initialize Face Sets" options in sculpt mode [2].

Discussion in T102962 has favored "Group" for a name for these
sockets rather than "Set", so that is used here.

[1]: https://docs.blender.org/manual/en/3.4/modeling/geometry_nodes/mesh/face_set_boundaries.html
[2]: https://docs.blender.org/manual/en/3.5/sculpt_paint/sculpting/editing/face_sets.html#initialize-face-sets