Nodes: Avoid updating topology cache when changing properties
Changing a property on a node or socket doesn't change the topology
of the node tree directly, so when only those tags are present, we don't
have to tag the topology cache dirty and update it.

This can save a few milliseconds per update when editing a large node tree.
Note that the topology cache still has to be recreated for the CoW node tree
though, so this doesn't resolve all of the unnecessary calculation.