Fix #101374: Enable edge panning for move operators in menus
This patch enables the edge panning for the two operators where there
was a mismatch between the behaviour when pressing G and the "Move"
operator exposed via the menu and operator search.

Commit 19da434e9cc0 introduced a view2d_edge_pan property which was
activated for the translate operators invoked when pressing G in the
default keymap.
Since the property is not enabled by default, to avoid it getting in the
way in some cases, the "Move" operators exposed via the menus didn't
have edge panning. This lead to inconsistent behavior between moving via
the shortcut or via the menu.