UI: Change Behavior for Multi-Select Enums
Changes:
- Remove the need to hold `Shift`.
- Hold `Ctrl` to deselect all options except the one clicked on.

Many enums in Blender are able to have multiple items selected, but
this feature isn't exposed very well.

It's not clear which enums support this, so you have to try by pressing
shift when clicking on them.

And a lot of the time it's more valuable to have multiple items
selected, but it's harder because of the modifier key.