Grease Pencil: Set Layers' 'Use Lights' Default to False
The default ‘Use Lights’ setting for new Layers is unwanted by and confusing to many Grease Pencil users. Since Grease Pencil is used primarily for non-photorealistic rendering, the assumption should be for the user to get the lines and fill colors they draw: then to add realistic lighting to it if desired. For new users, the colors are darker than drawn because ‘Use Lights’ is enabled by default which is very confusing.

https://blender.community/c/rightclickselect/qrhbbc/
https://blender.stackexchange.com/q/230073/81083

This patch disables ‘Use Lights’ by default so that experienced Grease Pencil users may enable this setting on a per-layer basis when desired.