Online help: Improve docs search time
Matching the rna id's to the search pattern is slow because
of the function `fnmatchcase`. This patch first checks if the pattern
without the last character contains an asterisk, and if not it uses
`startswith` instead of `fnmatchcase`.
The improvement in search time is from around 400ms average before
which is quite noticable, to 1ms average after.