Fix retopology overlay in orthographic view
After a few hours of trial and error, I managed to figure out the right way to calculate the bias value for orthographic view.
Also it turns out bias should divided by W instead of Z in the shaders; this fixes orthographic view, and some minor artifacts in perspective view.