Asset Brwoser: Add import settings popover, move import method into it
Replaces the current import method menu in the middle of the Asset
Browser header with a "Import Settings" popover. The import method is
located there now, because it doesn't need to be as prominent as it used
to be, especially with #104686. We can also add further settings to it,
such as a collection instance option.

Part of #104686.