Fix #104730: Make first visible CustomData layer active
When creating a CustomData layer with a name not starting with a dot
while all existing layers of the same type have names which do start
with a dot make the new layer the active/render/mask/clone layer.

It can be confusing if the render layers is a hidden/anonymous layer.