Fix #104810: Appending a camera does not pull in background movie clip
This was broken even before 0649e6371629 and was always expanding the
`Image`, not the movie clip (even if the source was set to
`CAM_BGIMG_SOURCE_MOVIE`).

Now the rule here seems to be to always expand unconditionally, so
remove checking the source and always expand image and movie clip.