Fix #104358: Crash with broken writing to freed evaluated mesh
For some reason that I have no idea about, the "Fix Deforms" operator
retrieved the evaluated mesh, wrote to its vertex group data, undid
that write, and then freed the evaluated mesh, and then tried to do
that again, resulting in a use-after free.

This operator is a huge mess and I have no idea what its goal is, but
this change fixes the crash by simply removing the no-op writing.
Ideally the entire operator should be rewritten or removed,
since the code is suspect in many other ways.