UI: Properly cancel dragging on Esc & right-click
Previously Escape would cancel the dragging, but still pass on the event
to other handlers, which could trigger other behavior. For example
cancelling dragging a file in a file browser dialog would cancel the
file browser operation and close the window. Right-click didn't cancel
anything even though we usually use both Escape and right-click as a way
to cancel the operation.

Now both Escape and right-click both cancel the dragging and the event
is not passed on further.