VSE: Add sound strip retiming support
This patch contains changes needed for retiming sound strips. This is a part of #104544

----

Deviation from #100337 is, that changes in audaspace are not separated.
Versioning for old files is not present, because resulting data is unusable (whole animation is baked, so changing retiming is basically impossible). Versioning also results in precision issues which results in non uniform sound pitch in some cases. And finally calculating strip offset correction is non-trivial (that is least of an issue, but worth mentioning still). So I would advocate to not convert old data to current implementation. I think, that supporting smooth speed transitions would allow for more efficient versioning code.