AxisAngle: Refactor to use different angle representation
This make use of template to change how the angle is stored.
This avoid duplicating all the constructors.

The downside is that the setup cost is a bit hidden behind the AxisAngle
type.

The AngleSinCos type could be used in other cases such as 2D rotation.

Naming convention is up to debate.