Add Grease Pencil Layer Context Menu to Dopesheet Channel Context Menu
The Grease Pencil Layer Context Menu is currently located inconveniently in the user interface. When wanting to select and act on specific layers, the context menu makes the most sense in the Dopesheet Channel Context Menu: it is closest to the layers and also the most obvious location to a user.