Fix Light Tree option causing wrong renders on macOS AMD
The light tree itself is disabled on the AMD GPUs due to a compiler issue.

There are couple of places where this was not fully checked:

- The `light_sample` function in the kernel.
- The light threshold during synchronization

The former one is solved as easy as just adding an ifdef block.

The latter one is solved by delaying the threshold assignment for
later on. There is some duplication of the the use_light_tree assignment
for the `kintegrator->use_light_tree` (happens in the integrator.cpp
and light.cpp). Not sure what the initial intent for the source of truth
was, but it seems safer than having two non-match assignments in different
places.