Unflattens the vertex and index buffers
Vertex buffers and index buffers are now stored as a set of vertices
and the corresponding triangle indices that make up the meshes.
Previously the vertex buffers were flattened meaning the vertices
were repeated for each triangle storing 3 vertices for each triangle.
Storing the meshes this way reduces the amount of space required
to store the mesh and thus reduces the cost of transferring the
data to the GPU and improves caching. The patch also stores the
vertices as packed_float3 and the index buffers are packed_uint3.